rename and inline
